Update opensim-core commit to 4.0

+++ Muscollo/Muscollo/MucoParameter.cpp
@@ -132,9 +132,9 @@ void MucoParameter::printDescription(std::ostream& stream) const {
     stream << ". model property name: " << getPropertyName();
     stream << ". component paths: ";
     const std::vector<std::string> componentPaths = getComponentPaths();
-    for (int i = 0; i < componentPaths.size(); ++i) {
+    for (int i = 0; i < (int)componentPaths.size(); ++i) {
         stream << componentPaths[i];
-        if (i < componentPaths.size()-1) {
+        if (i < (int)componentPaths.size()-1) {
            stream << ", ";
         } else {
            stream << ". ";
